% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renderengine.R
\name{Renderengine}
\alias{Renderengine}
\title{Renderengine}
\usage{
Renderengine(PATH, cores)
}
\arguments{
\item{PATH}{path, "." current directory}

\item{cores}{change number of cores to be used}
}
\value{
blbla
}
\description{
This functions renders rmarkdown files.
}
